Inline *c.c files into the headers, declare functions to be static
