+++ dune-project
@@ -1,0 +1,1 @@
+(lang dune 1.10)
